move callback functions to ubertooth_callback.c
