+++ host/ubertooth-tools/src/ubertooth-ego.c
@@ -20,6 +20,7 @@
  */
 
 #include "ubertooth.h"
+#include "ubertooth_callback.h"
 #include <ctype.h>
 #include <err.h>
 #include <getopt.h>
